requirements.txt
httpx>=0.27.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
numpy>=1.24.0
faiss-gpu>=1.7.4
tree-sitter>=0.22.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-cpp>=0.23.0
tree-sitter-c>=0.23.0
tree-sitter-c-sharp>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-rust>=0.23.0
tree-sitter-ruby>=0.23.0
tree-sitter-php>=0.23.0
tree-sitter-kotlin>=0.23.0
